submit C.cpp to 'C - 3 Steps' (code-festival-2017-qualb) [C++14 (GCC 5.4.1)]

+++ 2017/1008_CF2017-qualB/C.cpp
@@ -34,7 +34,7 @@ typedef long long ll;
 typedef tuple<int, int, int> state; // now, from, clr
 typedef tuple<int, int> cond; // now, depth
 
-int N;
+ll N;
 int M;
 vector<int> V[100010];
